Separate classes and struct into new files.
This had to be done. The general feel of the Dungen class was crowded and illogical. With all the structs and classes put into separate files, this feeling is now mainly gone.

Lots of stuff still has to be done, but I've created two TO-DO lists within the main Dungen Class.